+++ CHANGELOG.rst
@@ -1,6 +1,15 @@
 Changelog
 *********
 
+0.37.0 (2023-09-09)
+-------------------
+
+* :issue:`67`: Fix bug in :meth:`pymaf.MafFrame.plot_waterfall` method where ``count=1`` was causing color mismatch.
+* Add new submodule ``pychip``.
+* Add new method :meth:`common.reverse_complement`.
+* Fix bug in :meth:`common.extract_sequence` method where a long DNA sequence output was truncated.
+* :issue:`68`: Refresh the variant consequences database from Ensembl VEP. The database's latest update was on May 31, 2021.
+
 0.36.0 (2022-08-12)
 -------------------